ObliQ/cvar-vqe/requirements.txt
appnope==0.1.4
asttokens==3.0.0
certifi==2025.10.5
charset-normalizer==3.4.4
comm==0.2.3
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
drawsvg==2.4.0
exceptiongroup==1.3.0
executing==2.2.1
exqalibur==1.0.0
fonttools==4.60.1
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.31.0
ipython==8.18.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.7
latexcodec==3.0.1
matplotlib==3.9.4
matplotlib-inline==0.2.1
mpmath==1.3.0
multipledispatch==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==2.0.2
packaging==25.0
parso==0.8.5
perceval-quandela==1.0.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
protobuf==6.33.0
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pyzmq==27.1.0
requests==2.32.5
scipy==1.13.1
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.14
zipp==3.23.0
